+++ gemsiv/arcball/BallMath.h
@@ -3,7 +3,7 @@
 #define _H_BallMath
 #include "BallAux.h"
 
-HVect MouseOnSphere(HVect mouse, HVect ballCenter, double ballRadius);
+HVect MouseOnSphere(HVect mouse, HVect ballCenter, float ballRadius);
 HVect ConstrainToAxis(HVect loose, HVect axis);
 int NearestConstraintAxis(HVect loose, HVect *axes, int nAxes);
 Quat Qt_FromBallPoints(HVect from, HVect to);
