+++ gemsiv/vert_norm/test.c
@@ -11,7 +11,7 @@
 
 /* normalizes the input vector and returns it */
 Vector3 *V3Normalize(Vector3 *v) {
-	double len = sqrt(V3Dot(v, v));
+	float len = sqrtf(V3Dot(v, v));
 	if (len != 0.0) { v->x /= len;  v->y /= len; v->z /= len; }
 	return(v);
 }
@@ -23,20 +23,20 @@ Vector3 *V3Add(Vector3 *a, Vector3 *b, Vector3 *c) {
 }
 
 /* return the dot product of vectors a and b */
-double V3Dot(Vector3 *a, Vector3 *b) {
+float V3Dot(Vector3 *a, Vector3 *b) {
 	return((a->x*b->x)+(a->y*b->y)+(a->z*b->z));
 }
 #endif
 
 /* z=f(x,y) */
-double fofxy(double x, double y) {
-	double h;
-	h = 2.0 * (0.5 - x); if (h < 0) h = -h; h = h * y;
+float fofxy(float x, float y) {
+	float h;
+	h = 2.f * (0.5f - x); if (h < 0) h = -h; h = h * y;
 	return(h);
 }
 
 
-void addTriangles(double lx, double ly, double hx, double hy,
+void addTriangles(float lx, float ly, float hx, float hy,
 	Point3 *vlist, Smooth smooth) {
 		Point3 *p = vlist;
 		/* make the first triangle */
@@ -52,7 +52,7 @@ void addTriangles(double lx, double ly, double hx, double hy,
 		includePolygon(3, vlist, smooth, NULL);  /* add the polygon */
 }
 
-void addQuadrilateral(double lx, double ly, double hx, double hy,
+void addQuadrilateral(float lx, float ly, float hx, float hy,
 	Point3 *vlist, Smooth smooth) {
 		Point3 *p = vlist;
 		p->x = lx;  p->y = ly;  p->z = fofxy(p->x, p->y); p++;
@@ -65,10 +65,10 @@ void addQuadrilateral(double lx, double ly, double hx, double hy,
 void buildMesh(Smooth smooth, int xres, int yres) {
 	int x, y;
 	Point3 *vlist;
-	double dx, dy, lx, ly, hx, hy;
+	float dx, dy, lx, ly, hx, hy;
 	vlist = NEWA(struct Point3Struct, 4);
-	dx = 1.0/((double)(xres));
-	dy = 1.0/((double)(yres));
+	dx = 1.f/((float)(xres));
+	dy = 1.f/((float)(yres));
 	for (y=0; y<yres; y++) {
 		ly = y * dy;
 		hy = (y+1) * dy;
@@ -89,7 +89,7 @@ void savePolys(Smooth smooth) {
 	printf("NQUAD\n");	/* header for point/normal format */
 	while (poly != NULL) {
 		for (i=0; i<4; i++) {
-			k = i;	  /* we always write 4 points so double 3rd triangle vertex */
+			k = i;	  /* we always write 4 points so float 3rd triangle vertex */
 			if (i >= poly->numVerts) k = poly->numVerts-1;
 			v = &(poly->vertices[k]);
 			n = &(poly->normals[k]);
